+++ packages/leancode_cubit_utils_cqrs/lib/src/request/use_query_cubit.dart
@@ -10,88 +10,42 @@ class SimpleQueryCubit<TOut> extends QueryCubit<TOut, TOut> {
   SimpleQueryCubit(
     super.loggerTag,
     this._customRequest, {
+    EmptyChecker<TOut>? isEmpty,
     super.requestMode,
-  });
+  }) : _isEmpty = isEmpty;
 
   /// The request to be executed.
   final Request<QueryResult<TOut>> _customRequest;
+
+  /// The function to check if the data is empty.
+  final EmptyChecker<TOut>? _isEmpty;
 
   @override
   Future<QueryResult<TOut>> request() => _customRequest();
 
   @override
   TOut map(TOut data) => data;
+
+  @override
+  bool isEmpty(TOut data) => _isEmpty?.call(data) ?? false;
 }
 
 /// Provides a [QueryCubit] specialized for [QueryResult] that is automatically disposed without having
 /// to use BlocProvider and does not require any arguments. It is a wrapper of [useBloc] that creates a [SimpleQueryCubit].
 SimpleQueryCubit<TOut> useQueryCubit<TOut>(
   Request<QueryResult<TOut>> request, {
-  String loggerTag = 'SimpleQueryCubit',
+  String loggerTag = 'SimpleQueryWithEmptyCubit',
   RequestMode? requestMode,
   bool callOnCreate = true,
   List<Object?> keys = const [],
+  EmptyChecker<TOut>? isEmpty,
 }) {
   return useBloc(
     () {
       final cubit = SimpleQueryCubit<TOut>(
         loggerTag,
         request,
-        requestMode: requestMode,
-      );
-      if (callOnCreate) {
-        cubit.run();
-      }
-      return cubit;
-    },
-    keys: keys,
-  );
-}
-
-/// Simplified implementation of [QueryCubit] created in order to be used by [useQueryCubit].
-/// Differ from [SimpleQueryCubit] because it uses a custom function to check if the data is empty.
-class SimpleQueryWithEmptyCubit<TOut> extends QueryCubit<TOut, TOut> {
-  /// Creates a new [SimpleQueryWithEmptyCubit].
-  SimpleQueryWithEmptyCubit(
-    super.loggerTag,
-    this._customRequest,
-    this._isEmpty, {
-    super.requestMode,
-  });
-
-  /// The request to be executed.
-  final Request<QueryResult<TOut>> _customRequest;
-
-  /// The function to check if the data is empty.
-  final EmptyChecker<TOut> _isEmpty;
-
-  @override
-  Future<QueryResult<TOut>> request() => _customRequest();
-
-  @override
-  TOut map(TOut data) => data;
-
-  @override
-  bool isEmpty(TOut data) => _isEmpty(data);
-}
-
-/// Provides a [QueryCubit] specialized for [QueryResult] that is automatically disposed without having
-/// to use BlocProvider and does not require any arguments. It is a wrapper of [useBloc] that creates a [SimpleQueryWithEmptyCubit].
-/// Differ from [useQueryCubit] because it uses a custom function to check if the data is empty.
-SimpleQueryWithEmptyCubit<TOut> useQueryWithEmptyCubit<TOut>(
-  Request<QueryResult<TOut>> request,
-  EmptyChecker<TOut> isEmpty, {
-  String loggerTag = 'SimpleQueryWithEmptyCubit',
-  RequestMode? requestMode,
-  bool callOnCreate = true,
-  List<Object?> keys = const [],
-}) {
-  return useBloc(
-    () {
-      final cubit = SimpleQueryWithEmptyCubit<TOut>(
-        loggerTag,
-        request,
-        isEmpty,
+        isEmpty: isEmpty,
         requestMode: requestMode,
       );
       if (callOnCreate) {
@@ -110,8 +64,8 @@ class SimpleArgsQueryCubit<TArgs, TOut>
   SimpleArgsQueryCubit(
     super.loggerTag,
     this._customRequest, {
-    super.requestMode,
     EmptyChecker<TOut>? isEmpty,
+    super.requestMode,
   }) : _isEmpty = isEmpty;
 
   /// The request to be executed.
@@ -127,7 +81,7 @@ class SimpleArgsQueryCubit<TArgs, TOut>
   TOut map(TOut data) => data;
 
   @override
-  bool isEmpty(TOut data) => _isEmpty?.call(data) ?? data == null;
+  bool isEmpty(TOut data) => _isEmpty?.call(data) ?? false;
 }
 
 /// Provides a [ArgsQueryCubit] specialized for [QueryResult] that is automatically disposed without having
@@ -136,62 +90,14 @@ SimpleArgsQueryCubit<TArgs, TOut> useArgsQueryCubit<TArgs, TOut>(
   ArgsRequest<TArgs, QueryResult<TOut>> request, {
   String loggerTag = 'SimpleArgsQueryCubit',
   RequestMode? requestMode,
+  EmptyChecker<TOut>? isEmpty,
   List<Object?> keys = const [],
 }) {
   return useBloc(
     () => SimpleArgsQueryCubit<TArgs, TOut>(
       loggerTag,
       request,
-      requestMode: requestMode,
-    ),
-    keys: keys,
-  );
-}
-
-/// Simplified implementation of [ArgsQueryCubit] created in order to be used by [useArgsQueryCubit].
-/// Differ from [SimpleArgsQueryCubit] because it uses a custom function to check if the data is empty.
-class SimpleArgsQueryWithEmptyCubit<TArgs, TOut>
-    extends ArgsQueryCubit<TArgs, TOut, TOut> {
-  /// Creates a new [SimpleArgsQueryWithEmptyCubit].
-  SimpleArgsQueryWithEmptyCubit(
-    super.loggerTag,
-    this._customRequest,
-    this._isEmpty, {
-    super.requestMode,
-  });
-
-  /// The request to be executed.
-  final ArgsRequest<TArgs, QueryResult<TOut>> _customRequest;
-
-  /// The function to check if the data is empty.
-  final EmptyChecker<TOut> _isEmpty;
-
-  @override
-  Future<QueryResult<TOut>> request(TArgs args) => _customRequest(args);
-
-  @override
-  TOut map(TOut data) => data;
-
-  @override
-  bool isEmpty(TOut data) => _isEmpty(data);
-}
-
-/// Provides a [ArgsQueryCubit] specialized for [QueryResult] that is automatically disposed without having
-/// to use BlocProvider and requires arguments. It is a wrapper of [useBloc] that creates a [SimpleArgsQueryWithEmptyCubit].
-/// Differ from [useArgsQueryCubit] because it uses a custom function to check if the data is empty.
-SimpleArgsQueryWithEmptyCubit<TArgs, TOut>
-    useArgsQueryWithEmptyCubit<TArgs, TOut>(
-  ArgsRequest<TArgs, QueryResult<TOut>> request,
-  EmptyChecker<TOut> isEmpty, {
-  String loggerTag = 'SimpleArgsQueryWithEmptyCubit',
-  RequestMode? requestMode,
-  List<Object?> keys = const [],
-}) {
-  return useBloc(
-    () => SimpleArgsQueryWithEmptyCubit<TArgs, TOut>(
-      loggerTag,
-      request,
-      isEmpty,
+      isEmpty: isEmpty,
       requestMode: requestMode,
     ),
     keys: keys,